Make EventModel more QML friendly
EventModel can be made usable in QML by declaring properties and
offering the same data as in columns through named roles, without
creating a dependency on QtDeclarative or changing behavior in any
other way.

These changes couldn't be reasonably done in a wrapper, because
they'd have to be repeated for every subclass of EventModel.