Add Group to the declarative API
libcommhistory's Group is a non-QObject and never receives updates,
so it's unusable for QML. GroupObject is a wrapper, created via a
GroupProxyModel instance, which exposes the group's properties to QML
and watches for updates or removal using the model's signals, which
are the only source for that information.

GroupProxyModel is a QIdentityProxyModel wrapping a libcommhistory
GroupModel to expose its API in a QML-friendly way. Currently, it
requires that the user set sourceModel manually, but internal creation
can be added as an option later.