[libcommhistory-declarative] Add a method to create basic outgoing message events
This allows clients to create outgoing message events that will be
immediately saved and visible in the UI regardless of actual backend
state, because it may take some time before it's possible to actually
send the message.

The returned event ID can be sent in the telepathy message
x-commhistory-event-id header to prevent commhistory-daemon from
duplicating the event.

Currently, this only supports plaintext IM and SMS messages with a
single target.