Implement early basic sqlite database code
Includes basic creation of the database, porting of the models. It's
possible to add and list groups, events, and calls in a basic way.

Most of the database behavior is stubbed out, and many things are
disabled or broken.

QStandardPaths is used for the database location, or a hardcoded
equivalent on Qt4.