Fix behavior of group startTime and endTime
The startTime and endTime of a group are actually the startTime and
endTime of its last event (which is confusing, but regardless). Query
them directly rather than storing them with the group.

The "last event" in a group is actually the event with the most recent
startTime, falling back to descending ID order. Fix the groups query to
obey those rules.

Also fixed the last event ID for groups with no events. It should be -1.