[libcommhistory] Add setEventStatus method to DeclarativeGroupManager
Technically, this has nothing to do with groups. Regardless, there is no
other sane place to put this with the current API, and it makes no sense
to create a more elaborate event API for the current scope of the
declarative plugin.

It couldn't be reliably done as part of the ConversationModel without
also having use of SingleEventModel as a fallback, and that would be
much less convenient for usage.