[libcommhistory] Increase the maximum digits used in minimized phone numbers
Although 7 digits is typically enough to differentiate phone numbers
in their local forms, an eighth digit is sometimes required to
differentiate between a local number and the voicemail service for
that number.

Also, remove the check to see if the maximum value is overridden by
user configuration file.  This could lead to confusing discrepancies
between commhistory and contact number resolution, and should eventually
be replaced by a system-wide configuration setting.