[libcommhistory] Assign in-model-only events valid IDs
Before migration to SQLite, events stored in-model-only used to be
assigned an artificial ID. This changed with 91ab2f8 (Implement early
basic sqlite database code) so that such events are set an invalid ID =
-1 (search for use of tracker()->nextEventId() in 91ab2f8 diff). This
broke this in-model-only events handling and it was first noticed in
case of ClassZeroSMSEvent. An attempt to fix this with e3a1901 (Fix
exporting and behavior of ClassZeroSMSModel) helped to discover the
actual issue as it actually inverted its effect instead of fixing it:
every event added in-model-only is inserted twice because
findEventRecursive() modified with e3a1901 never success finding event
with an invalid id and so it gets reinserted when self-emitted
eventsAdded() D-Bus signal is received (detected by
ConversationModelTest::getEvent()).

This commit reintroduces assigning the in-model-only events artificial
ids.