[commhistory-tool] Add import-json -relativeDate option
Specifying '-relativeDate 20131204' to import-json will offset
imported events so that events on December 4, 2013 will have today's
date, and all other dates will be adjusted accordingly. Times for
imported events are not modified, which can result in events a few hours
in the future on the current day.

This is intended to help when importing predefined content.