[libcommhistory] Add ContactResolver class
ContactResolver resolves contacts matching a list of events, and emits a
signal when all events have been resolved (successfully or not). It's
currently built on ContactListener, but could be made more efficient by
using SeasideCache directly later (e.g by handling results from the
cache instantly, and reducing address comparisons). It is intended to
eventually evolve into an API that can replace the bad ContactListener
API entirely.

ContactResolver maintains a cache of lookups to contact results,
intended to provide a fast path when resolving many events using the
same address (which may not exactly match any of a contact's addresses).