[libcommhistory] Use localUid to determine when remote UIDs are phone numbers
Previously, a remoteUid was compared as a phone number (i.e. normalized
and minimized) if it passed validation as a phone number. That logic was
inconsistent with contact resolution and flawed for some
weirdly-formatted numbers, leading in the worst case to hung contact
resolve tasks.

Instead, use the localUid to determine whether remote UIDs are compared
as phone numbers, and normalize them without validation.

This involves API changes to several utility functions, but commhistory
does not currently promise any API stability.