[tests] Skip test for RecentContactsModel::requiredProperty
Getting this property to work again was too much work considering
we don't use it. One detail making it hard to fix it was that
we can get addressResolved() for a contact before the contact
has all of its addresses, so prependEvents() is too early to
decide to discard an event.

Removed some time-and-memory consuming code that used to be related.
Added a warning if someone tries to use the property..

This functionality was lost in commit 096972c9